Worktree cleanup: Baritone port audit fixes + Webcore UI (UNTESTED)
UNTESTED — committed to clean up the accumulated worktree ahead of a
testing pass. Builds clean (Core + Baritone) and 31/31 unit tests pass,
but none of this has been validated against a live server yet.

Baritone port (1:1 parity audit vs Java 1.21.11):
- Movement cost/state rewrites (Parkour, Ascend, Descend, Traverse,
  Diagonal, Pillar, Downward, Fall) + magma/frost-walker/depth-strider.
- Full mining-cost model: generated hardness table (TOOL1), enchantments
  (Efficiency/SilkTouch), tool selection, potion effects (Haste/Fatigue).
- S5 VoxelShape wiring (walkability, placement shape, isInWall/H6).
- Segment plan-ahead (PB1), path sanity + cutoffs (PA1/PA2).
- Aim: mouse-grid quantization (AIM1), ForkableRandom jitter (AIM5/6),
  tick-order fix reviving smoothing/reset (AIM4 — needs runtime check).
- Doors/fence-gates (T2) + vanilla 4-tick rightClickDelay.
- disconnectOnArrival (PB2), lag-rewind parity (PE-H5), bitmask (S3).

New Core infra: EnchantmentHelper, BlockDefinition/BlockHardness,
ClientState enchantment/mob-effect registries, Entity active-effects,
mob-effect packet handling.

Webcore + auth: multi-account manager, Accounts page, movement panel,
rich container rendering, panic switch, humanizer, UI facelift.

Note: StaticFiles/ is gitignored (regenerated locally); the new
block_hardness.json is NOT yet wired into generate-all.bat, so a fresh
checkout needs it regenerated before TOOL1 hardness is accurate.

+++ MinecraftProtoNet.Baritone/Pathfinding/Goals/GoalAxis.cs
@@ -44,7 +44,8 @@ public class GoalAxis : Goal
         int lng = Math.Max(x, z);
         int diff = lng - shrt;
 
-        double flatAxisDistance = Math.Min(x, Math.Min(z, (int)(diff * Sqrt2Over2)));
+        // Reference: GoalAxis.java:40 - kept as double (no int cast) to match the Java heuristic exactly
+        double flatAxisDistance = Math.Min((double)x, Math.Min((double)z, diff * Sqrt2Over2));
 
         return flatAxisDistance * BaritoneAPI.GetSettings().CostHeuristic.Value + GoalYLevel.Calculate(BaritoneAPI.GetSettings().AxisHeight.Value, y);
     }
